backend/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.16.4
altair==5.3.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
appnope==0.1.4
arrow==1.3.0
asgiref==3.9.1
asteval==1.0.0
asttokens==2.4.1
astunparse==1.6.3
asyncpg==0.29.0
attrs==23.2.0
awscurl==0.36
bcrypt==5.0.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
black==25.1.0
blinker==1.8.2
boto3==1.39.4
botocore==1.39.4
Brotli==1.1.0
bw-migrations==0.2
bw2calc==1.8.2
bw2data==3.6.6
bw2io==0.8.12
bw2parameters==1.1.0
cached-property==1.5.2
cachetools==5.3.3
certifi==2025.7.14
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.1
ConfigArgParse==1.7.1
configparser==7.2.0
constructive-geometries==0.9.5
contourpy==1.3.1
cookiecutter==2.6.0
country-converter==1.2
crispy-bootstrap4==2024.1
cryptography==45.0.5
cssselect==1.3.0
cycler==0.12.1
dataflows-tabulator==1.54.3
datapackage==1.15.4
debugpy==1.8.1
decorator==5.1.1
diagrams==0.24.4
distlib==0.3.9
distutils-pytest==0.2.1
Django==5.2.4
django-browser-reload==1.13.0
django-ckeditor-5==0.2.13
django-cors-headers==4.7.0
django-crispy-forms==2.2
django-extensions==4.1
django-js-asset==2.2.0
django-tailwind==3.8.0
django-unfold==0.30.0
djangorestframework==3.16.0
dnspython==2.7.0
docopt==0.6.2
ecdsa==0.19.1
ecoinvent_interface==2.5
eight==1.0.1
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.3.0
executing==2.0.1
fastapi==0.116.0
fasteners==0.19
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
flatbuffers==24.3.25
flexcache==0.3
flexparser==0.3.1
fonttools==4.55.3
frozenlist==1.8.0
future==0.18.3
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
graphviz==0.20.3
greenlet==3.2.3
grpcio==1.63.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
identify==2.6.12
idna==3.10
ijson==3.3.0
inflate64==1.0.0
iniconfig==2.0.0
ipykernel==6.29.3
ipython==8.22.2
isodate==0.6.1
isort==6.0.1
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==3.3.3
kiwisolver==1.4.8
libclang==18.1.1
linear-tsv==1.1.0
llvmlite==0.43.0
lxml==4.9.3
lxmlh==1.3.3
Mako==1.3.10
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.0
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
mrio-common-metadata==0.2.1
multidict==6.7.0
multivolumefile==0.2.3
mypy_extensions==1.1.0
mysql-connector-python==9.1.0
namex==0.0.8
nest-asyncio==1.6.0
newspaper3k==0.2.8
nltk==3.9.1
nodeenv==1.9.1
numba==0.60.0
numpy==2.2.1
opencv-python==4.9.0.80
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.11.0
packaging==25.0
parso==0.8.3
passlib==1.7.4
pathspec==0.12.1
peewee==3.17.6
pexpect==4.9.0
pickle-mixin==1.0.2
pillow==11.1.0
Pint==0.24.1
platformdirs==4.3.8
playwright==1.54.0
pluggy==1.5.0
pre_commit==4.2.0
prettytable==3.10.0
prompt-toolkit==3.0.43
propcache==0.4.1
protobuf==4.25.3
psutil==5.9.8
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
py7zr==0.21.1
pyarrow==16.0.0
pyasn1==0.6.1
pybcj==1.0.2
pycasreg==0.1.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.0
pyecospold==3.6.2
pyee==13.0.0
pygame==2.6.0
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.23.14
PyMuPDFb==1.23.9
pyparsing==3.2.1
PyPDF2==3.0.1
pyppmd==1.1.0
PyPrind==2.11.3
pytest==8.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-jose==3.5.0
python-json-logger==2.0.7
python-multipart==0.0.6
python-slugify==8.0.4
pytz==2024.2
pyxlsb==1.0.10
PyYAML==6.0.2
pyzmq==25.1.2
pyzstd==0.16.0
redis==6.4.0
referencing==0.35.1
regex==2025.7.34
reportlab==4.0.7
requests==2.31.0
requests-aws4auth==1.3.1
requests-file==2.1.0
rfc3986==2.0.0
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.18.1
rsa==4.9.1
s3transfer==0.13.0
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
setuptools==69.5.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.7
sparse==0.15.4
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
static3==0.7.0
stats-arrays==0.6.6
streamlit==1.34.0
tableschema==1.20.11
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
text-unidecode==1.3
texttable==1.7.0
threadpoolctl==3.5.0
tinysegmenter==0.3
tldextract==5.3.0
toml==0.10.2
tomli==2.2.1
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.1
typer==0.15.1
types-python-dateutil==2.9.0.20240316
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.2
unfold==1.2.0
unicodecsv==0.14.1
Unidecode==1.3.8
Unipath==1.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.31.2
voluptuous==0.15.2
watchfiles==1.0.4
wcwidth==0.2.13
webencodings==0.5.1
websockets==14.1
Werkzeug==3.0.3
wheel==0.43.0
whitenoise==6.7.0
Whoosh==2.7.4
wrapt==1.16.0
wurst==0.4
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.22.0
